v1.7.1-release

+++ plugin/mysql/src/mysql_sql.go
@@ -2,6 +2,7 @@ package src
 
 import (
 	pluginDriver "github.com/brokercap/Bifrost/plugin/driver"
+	"regexp"
 	"strings"
 	"log"
 )
@@ -34,6 +35,14 @@ func (This *Conn) getAutoTableSqlSchemaAndTable(name string,DefaultSchemaName st
 	return
 }
 
+// 将sql 里 /* */ 注释内容给去掉
+// 感谢 @zeroone2005 正则表达式提供支持
+var replaceSqlNotesReq = regexp.MustCompile(`/\*(.*?)\*/`)
+func (This *Conn)  TransferNotes2Space(sql string) string {
+	sql = replaceSqlNotesReq.ReplaceAllString(sql, "")
+	return sql
+}
+
 //去除连续的两个空格
 func (This *Conn) ReplaceTwoReplace(sql string) string {
 	for {
@@ -45,8 +54,15 @@ func (This *Conn) ReplaceTwoReplace(sql string) string {
 	}
 }
 
-func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSql string) {
-	Query := strings.Trim(data.Query," ")
+func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSqlArr []string) {
+	// 优先判断是否 DML 语句
+	newSqlArr = This.TranferDMLSql(data)
+	if len(newSqlArr) > 0 {
+		return
+	}
+	var newSql string
+	var Query = strings.Trim(data.Query," ")
+	Query = This.TransferNotes2Space(Query)
 	// 变量 sql 是就不用拼接最后的  可执行 sql的，所以可以全部转成大写
 	sql := strings.ToUpper(Query)
 	// 防止连续多空格
@@ -64,18 +80,25 @@ func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSql str
 		SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(sqlArr[2],data.SchemaName)
 		sqlArr[2] = "`" + SchemaName + "`.`" + TableName +"`"
 		newSql = strings.Join(sqlArr, " ")
+		newSqlArr = append(newSqlArr,newSql)
 		goto End
 	}
 
 	// TRUNCATE TABLE tableName
-	if strings.Index(sql,"TRUNCATE TABLE") == 0 {
+	if strings.Index(sql,"TRUNCATE") == 0 {
 		Query = This.ReplaceTwoReplace(Query)
 		sqlArr := strings.Split(Query, " ")
-		var tableNameIndex = 2
+		var tableNameIndex int
+		if strings.Index(sql,"TRUNCATE TABLE ") == 0 {
+			tableNameIndex = 2
+		}else{
+			tableNameIndex = 1
+		}
 		SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(sqlArr[tableNameIndex],data.SchemaName)
 		var schemaAndTable = "`" + SchemaName + "`.`" + TableName +"`"
 		sqlArr[tableNameIndex] = schemaAndTable
 		newSql = strings.Join(sqlArr, " ")
+		newSqlArr = append(newSqlArr,newSql)
 		goto End
 	}
 
@@ -111,6 +134,7 @@ func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSql str
 			}
 			newSql = strings.Replace(Query,sqlArr[tableNameIndex],schemaAndTable,1)
 		}
+		newSqlArr = append(newSqlArr,newSql)
 		goto End
 	}
 
@@ -119,11 +143,12 @@ func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSql str
 		Query = This.ReplaceTwoReplace(Query)
 		sqlArr := strings.Split(Query, " ")
 		if strings.Index(sql,"IF NOT EXISTS") < 0 {
-			SchemaName = sqlArr[2]
+			sqlArr[1] = "DATABASE IF NOT EXISTS"
+			newSql = strings.Join(sqlArr, " ")
 		}else{
-			SchemaName = sqlArr[5]
+			newSql = Query
 		}
-		newSql = "CREATE DATABASE IF NOT EXISTS " + SchemaName + " ;"
+		newSqlArr = append(newSqlArr,newSql)
 		goto End
 	}
 
@@ -167,11 +192,20 @@ func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSql str
 			if t.From == "" || t.To == "" {
 				continue
 			}
-			if newSql == "" {
-				newSql += "RENAME TABLE " + t.From + " TO " + t.To
+			// TiDB 不支持  一条语句，多次 rename , 所以要分成多个rename 语句
+			if This.isTiDB {
+				newSql = "RENAME TABLE " + t.From + " TO " + t.To
+				newSqlArr = append(newSqlArr,newSql)
 			}else{
-				newSql += "," + t.From + " TO " + t.To
+				if newSql == "" {
+					newSql = "RENAME TABLE " + t.From + " TO " + t.To
+				}else{
+					newSql += "," + t.From + " TO " + t.To
+				}
 			}
+		}
+		if This.isTiDB == false && newSql != "" {
+			newSqlArr = append(newSqlArr,newSql)
 		}
 		goto End
 	}
@@ -191,6 +225,32 @@ func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSql str
 		}
 		sqlArr[tableNameIndex] = schemaAndTable
 		newSql = strings.Join(sqlArr, " ")
+		newSqlArr = append(newSqlArr,newSql)
+		goto End
+	}
+
+	// CREATE INDEX index_name ON table_name (column_name)
+	// CREATE UNIQUE INDEX index_name ON table_name (column_name)
+	if strings.Index(sql,"CREATE INDEX") == 0 || strings.Index(sql,"CREATE UNIQUE INDEX") == 0 {
+		Query = This.ReplaceTwoReplace(Query)
+		sqlArr := strings.Split(Query, " ")
+		var tableNameIndex = 4
+		if strings.Index(sql,"CREATE INDEX") != 0 {
+			tableNameIndex = 5
+		}
+		var schemaAndTable string
+		//CREATE INDEX indexName ON table(id int) 这种表名和( 相挨着的情况
+		if strings.Index(sqlArr[tableNameIndex],"(") > 0 {
+			tmpTableName := strings.Split(sqlArr[tableNameIndex], "(")[0]
+			SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(tmpTableName,data.SchemaName)
+			schemaAndTable = "`" + SchemaName + "`.`" + TableName +"`"
+			newSql = strings.Replace(Query,tmpTableName+"(",schemaAndTable+"(",1)
+		}else{
+			SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(sqlArr[tableNameIndex],data.SchemaName)
+			schemaAndTable = "`" + SchemaName + "`.`" + TableName +"`"
+			newSql = strings.Replace(Query,sqlArr[tableNameIndex],schemaAndTable,1)
+		}
+		newSqlArr = append(newSqlArr,newSql)
 		goto End
 	}
 
@@ -204,53 +264,71 @@ func (This *Conn) TranferQuerySql(data *pluginDriver.PluginDataType) (newSql str
 			SchemaName = sqlArr[4]
 		}
 		newSql = "DROP DATABASE IF EXISTS " + SchemaName + ";"
+		newSqlArr = append(newSqlArr,newSql)
 		goto End
-	}else{
-		// UPDATE Table
-		// INSERT INTO Table
-		// DELETE FROM Table
-		// REPLACE INTO Table
-		var tableNameIndex = 0
-		// insert,update,replace 字符串后第几个非空的字符串，才是第表名
-		var x = 1
-		switch sql[0:6] {
-		case "UPDATE":
-			break
-		case "INSERT","DELETE","REPLAC":
-			x = 2
-			break
-		default:
-			return
+	}
+End:
+	return
+}
+
+func (This *Conn) TranferDMLSql(data *pluginDriver.PluginDataType) (newSqlArr []string) {
+	var Query = strings.TrimLeft(data.Query," ")
+	var SchemaName,TableName string
+	// UPDATE Table
+	// INSERT INTO Table
+	// DELETE FROM Table
+	// REPLACE INTO Table
+	var tableNameIndex = 0
+	// insert,update,replace 字符串后第几个非空的字符串，才是第表名
+	var x = 1
+	switch strings.ToUpper(Query[0:6]) {
+	case "UPDATE":
+		break
+	case "INSERT","DELETE","REPLAC":
+		x = 2
+		break
+	default:
+		return
+	}
+	// 这里不能使用 ReplaceTwoReplace 将  两个空格转成一个空格再进行计算 ，因为实际 insert 或者 update 等的内容里是值有可能是 两个空格的内容
+	// 这里也不能用 TransferNotes2Space , 因为实际insert ,upadte 语句中，就包括 /* */ 内容
+	// 这里采用 遍历 的方式，找第一个或者第二个非空的字段串，当作是表名
+	sqlArr := strings.Split(Query, " ")
+	var tmpX = 0
+	var inNotes = false
+	for i := 1;i < len(sqlArr) ; i++ {
+		var tmp = strings.Trim(sqlArr[i]," ")
+		if tmp == "" || tmp == " " {
+			continue
 		}
-		// 这里不能使用 ReplaceTwoReplace 将  两个空格转成一个空格再进行计算 ，因为实际 insert 或者 update 等的内容里是值有可能是 两个空格的内容
-		// 这里采用 遍历 的方式，找第一个或者第二个非空的字段串，当作是表名
-		sqlArr := strings.Split(Query, " ")
-		var tmpX = 0
-		for i := 1;i < len(sqlArr) ; i++ {
-			var tmp = strings.Trim(sqlArr[i]," ")
-			if tmp == "" || tmp == " " {
-				continue
-			}
-			tmpX++
-			if tmpX == x {
-				tableNameIndex = i
-				break
-			}
+		if strings.Index(tmp,"/*") >= 0 {
+			inNotes = true
 		}
-		tmpTableName := sqlArr[tableNameIndex]
-		var schemaAndTable string
-		if strings.Index(tmpTableName,"(") > 0 {
-			tmpTableName = strings.Split(tmpTableName,"(")[0]
-			SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(tmpTableName,data.SchemaName)
-			schemaAndTable = "`" + SchemaName + "`.`" + TableName +"`"
-			newSql = strings.Replace(Query,tmpTableName+"(",schemaAndTable+"(",1)
-		}else{
-			SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(tmpTableName,data.SchemaName)
-			schemaAndTable = "`" + SchemaName + "`.`" + TableName +"`"
-			newSql = strings.Replace(Query,tmpTableName,schemaAndTable,1)
+		if inNotes {
+			if strings.Index(tmp,"*/") >=0 {
+				inNotes = false
+			}
+			continue
+		}
+		tmpX++
+		if tmpX == x {
+			tableNameIndex = i
+			break
 		}
 	}
-
-End:
+	tmpTableName := sqlArr[tableNameIndex]
+	var schemaAndTable string
+	var newSql string
+	if strings.Index(tmpTableName,"(") > 0 {
+		tmpTableName = strings.Split(tmpTableName,"(")[0]
+		SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(tmpTableName,data.SchemaName)
+		schemaAndTable = "`" + SchemaName + "`.`" + TableName +"`"
+		newSql = strings.Replace(Query,tmpTableName+"(",schemaAndTable+"(",1)
+	}else{
+		SchemaName,TableName = This.getAutoTableSqlSchemaAndTable(tmpTableName,data.SchemaName)
+		schemaAndTable = "`" + SchemaName + "`.`" + TableName +"`"
+		newSql = strings.Replace(Query,tmpTableName,schemaAndTable,1)
+	}
+	newSqlArr = append(newSqlArr,newSql)
 	return
 }
